+++ compiler/testData/codegen/boxInline/complex/lambdaInLambda.kt
@@ -1,5 +1,5 @@
 // FILE: 1.kt
-
+// WITH_RUNTIME
 package test
 
 public class Input(val s1: String, val s2: String) {