+++ idea/testData/debugger/tinyApp/src/evaluate/singleBreakpoint/nestedInlineArguments.kt
@@ -9,7 +9,7 @@ fun main(args: Array<String>) {
 inline fun <reified T: Bar> List<T>.foo(key: Int): T? {
     // EXPRESSION: it.i
     // RESULT: Unresolved reference: it
-    //Breakpoint! (lambdaOrdinal = 0)
+    //Breakpoint! (lambdaOrdinal = 1)
     return this.firstOrNull { it.i == key }
 }