+++ idea/testData/intentions/conventionNameCalls/replaceContains/containsFromJava.kt
@@ -1,0 +1,4 @@
+fun foo() {
+    val c = Container()
+    c.cont<caret>ains(1)
+}