+++ idea/tests/org/jetbrains/kotlin/idea/debugger/KotlinSteppingTestGenerated.java
@@ -478,9 +478,21 @@ public class KotlinSteppingTestGenerated extends AbstractKotlinSteppingTest {
             doCustomTest(fileName);
         }
 
+        @TestMetadata("smartStepIntoInsideLambda.kt")
+        public void testSmartStepIntoInsideLambda() throws Exception {
+            String fileName = KotlinTestUtils.navigationMetadata("idea/testData/debugger/tinyApp/src/stepping/custom/smartStepIntoInsideLambda.kt");
+            doCustomTest(fileName);
+        }
+
         @TestMetadata("smartStepIntoInterfaceImpl.kt")
         public void testSmartStepIntoInterfaceImpl() throws Exception {
             String fileName = KotlinTestUtils.navigationMetadata("idea/testData/debugger/tinyApp/src/stepping/custom/smartStepIntoInterfaceImpl.kt");
+            doCustomTest(fileName);
+        }
+
+        @TestMetadata("smartStepIntoSubClass.kt")
+        public void testSmartStepIntoSubClass() throws Exception {
+            String fileName = KotlinTestUtils.navigationMetadata("idea/testData/debugger/tinyApp/src/stepping/custom/smartStepIntoSubClass.kt");
             doCustomTest(fileName);
         }